+++ lib/radiobutton_group.dart
@@ -2,7 +2,7 @@
 // https://github.com/akserg/monomer
 // All rights reserved.  Please see the LICENSE.md file.
 
-library monomer_checkbox_group;
+library monomer_radiobutton_group;
 
 import 'dart:html';
 
@@ -14,13 +14,13 @@ import 'list_base.dart';
 import 'component.dart';
 
 /**
- *  CheckBoxGroup is inherited from ListBase, but by default it allows multiple 
- *  selection and automatically provides comma separated values as value property.
+ *  RadioButtonGroup is inherited from ListBase, but by default it allows one 
+ *  selection and automatically provides value property.
  */
-@CustomTag('m-checkbox-group')
-class CheckBoxGroup extends ListBase with Polymer, Observable, Component {
+@CustomTag('m-radiobutton-group')
+class RadioButtonGroup extends ListBase with Polymer, Observable, Component {
   
-  static final _logger = LoggerFactory.getLoggerFor(CheckBoxGroup);
+  static final _logger = LoggerFactory.getLoggerFor(RadioButtonGroup);
   
   /******************
    * Initialisation *
@@ -29,19 +29,19 @@ class CheckBoxGroup extends ListBase with Polymer, Observable, Component {
   /**
    * Default factory constructor.
    */
-  factory CheckBoxGroup() {
-    return new Element.tag('div', 'm-checkbox-group');
+  factory RadioButtonGroup() {
+    return new Element.tag('div', 'm-radiobutton-group');
   }
   
   /**
-   * Constructor instantiated by the DOM when a CheckBoxGroup element has been created.
+   * Constructor instantiated by the DOM when a RadioButtonGroup element has been created.
    */
-  CheckBoxGroup.created() : super.created() {
-    allowMultipleSelection = true;
+  RadioButtonGroup.created() : super.created() {
+    allowMultipleSelection = false;
     labelPath = 'label';
     valuePath = 'value';
     valueSeparator = ',';
-    itemRenderer = "span.m-checkbox";
+    itemRenderer = "span.m-radiobutton";
   }
 
   /**
@@ -62,8 +62,9 @@ class CheckBoxGroup extends ListBase with Polymer, Observable, Component {
    * ItemRenderer changed the state
    */
   void onItemRendererChange(CustomEvent event) {
-    _logger.debug('onItemRendererChange $event');
-    if (event.target is Component) {
+    _logger.debug('onItemRendererChange ${event.target} ${event.detail}');
+    // We triggering events comes only from selected controls
+    if (event.target is Component && event.detail as bool) {
       toggleSelection((event.target as Component).data);
     }
   }